# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Go Vue Template",
        "contact": {},
        "license": {
            "name": "Unknown"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/crud/accessLogs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a list for all entities of the AccessLog type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sort by id. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by path. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by request_body. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.request_body",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by request_headers. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.request_headers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by response_body. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.response_body",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by response_headers. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.response_headers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by response_code. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.response_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by processing_duration. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.processing_duration",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by request_method. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.request_method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by user_id. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by ip_address. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.ip_address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by created_at. Value should be ASC or DESC. eg: ?sort.created_at=DESC",
                        "name": "sort.created_at",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIAccessLog"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a new AccessLog",
                "parameters": [
                    {
                        "description": "Creation parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APICreateAccessLogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIAccessLog"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Soft deletes a range of accessLogs by their ids",
                "parameters": [
                    {
                        "description": "List of ids to delete",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelcrud.IDList"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Hard delete accessLog",
                        "name": "hardDelete",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.DeletedCount"
                        }
                    }
                }
            }
        },
        "/crud/accessLogs/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a single AccessLog entity by their id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "AccessLog id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIGetAccessLogsResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Updates a single AccessLog entity based on their id",
                "parameters": [
                    {
                        "description": "Update parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIUpdateAccessLogRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "AccessLog id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIAccessLog"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Soft deletes a single AccessLog entity based on their id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "AccessLog id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelcrud.APIAccessLog"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "modelcrud.APIAccessLog": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "description": "uuid",
                    "type": "string"
                },
                "ip_address": {
                    "description": "character varying",
                    "type": "string"
                },
                "path": {
                    "description": "character varying",
                    "type": "string"
                },
                "processing_duration": {
                    "description": "integer",
                    "type": "integer"
                },
                "request_body": {
                    "type": "string"
                },
                "request_headers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "request_method": {
                    "description": "character varying",
                    "type": "string"
                },
                "response_body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "response_code": {
                    "description": "integer",
                    "type": "integer"
                },
                "response_headers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "modelcrud.APICreateAccessLogRequest": {
            "type": "object",
            "properties": {
                "ip_address": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "processing_duration": {
                    "type": "integer"
                },
                "request_body": {
                    "type": "string"
                },
                "request_headers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "request_method": {
                    "type": "string"
                },
                "response_body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "response_code": {
                    "type": "integer"
                },
                "response_headers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "modelcrud.APIGetAccessLogsResponse": {
            "type": "object",
            "properties": {
                "access_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modelcrud.APIAccessLog"
                    }
                },
                "next_offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "modelcrud.APIUpdateAccessLogRequest": {
            "type": "object",
            "properties": {
                "ip_address": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "processing_duration": {
                    "type": "integer"
                },
                "request_body": {
                    "type": "string"
                },
                "request_headers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "request_method": {
                    "type": "string"
                },
                "response_body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "response_code": {
                    "type": "integer"
                },
                "response_headers": {
                    "type": "object",
                    "additionalProperties": true
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "modelcrud.DeletedCount": {
            "type": "object",
            "properties": {
                "deleted_count": {
                    "type": "integer"
                }
            }
        },
        "modelcrud.IDList": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}